oubli const

+++ ft_strncpy.c
@@ -6,13 +6,15 @@
 /*   By: acourtin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/02 15:30:19 by acourtin          #+#    #+#             */
-/*   Updated: 2017/11/12 19:06:57 by acourtin         ###   ########.fr       */
+/*   Updated: 2017/11/12 21:28:10 by acourtin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_strncpy(char *dest, const char *src, unsigned int n)
+#include "libft.h"
+
+char	*ft_strncpy(char *dest, const char *src, size_t n)
 {
-	unsigned int i;
+	unsigned long i;
 
 	i = 0;
 	while (src[i] != '\0' && i < n)
